@jberthold 2024-05-22
## KEVM performance run

| Test                                | 3863-llvm-term-cache time | master-0f3148308 time | (3863-llvm-term-cache/master-0f3148308) time 
|-------------------------------------|---------------------------|-----------------------|----------------------------------------------
| erc20/hkg/balanceOf-spec.k          | 35.76                     | 47.92                 | 0.7462437395659431                           
| mcd/end-cash-pass-rough-spec.k      | 187.89                    | 224.16                | 0.838195931477516                            
| mcd/pot-join-pass-rough-spec.k      | 168.3                     | 197.18                | 0.8535348412617912                           
| erc20/hkg/approve-spec.k            | 37.35                     | 42.92                 | 0.8702236719478099                           
| erc20/hkg/totalSupply-spec.k        | 35.08                     | 40.09                 | 0.8750311798453478                           
| erc20/hkg/transfer-failure-1-spec.k | 45.69                     | 52.1                  | 0.8769673704414587                           
| mcd/vat-addui-fail-rough-spec.k     | 56.21                     | 60.77                 | 0.9249629751522133                           
| TOTAL                               | 566.28                    | 665.14                | 0.8513696364675106      

`kontrol` performance test cannot be run at the moment.