@geo2a 2024-05-22
I've ran the Kontrol tests, and the speedup is there too:
| Test                                                                          | 3863-llvm-term-cache time | master-0f3148308 time | (3863-llvm-term-cache/master-0f3148308) time
|-------------------------------------------------------------------------------|---------------------------|-----------------------|----------------------------------------------
| HevmTests.prove_revert                                                        | 21.96                     | 23.59                 | 0.9309029249682069
| BytesTypeTest.test_bytes4(bytes4)                                             | 32.0                      | 34.33                 | 0.9321293329449462
| BlockParamsTest.testWarp(uint256)-trace_options2                              | 30.22                     | 32.4                  | 0.932716049382716
| ExpectRevertTest.testFail_ExpectRevert_failAndSuccess()                       | 108.37                    | 115.57                | 0.9377000951804102
| AssertTest.testFail_assert_true()                                             | 49.13                     | 51.79                 | 0.9486387333462059
| AssumeTest.test_assume_staticCall(bool)                                       | 52.13                     | 54.51                 | 0.9563382865529262
| StoreTest.testStoreLoad()                                                     | 60.28                     | 62.92                 | 0.958041958041958
| StoreTest.testGasLoadColdVM()                                                 | 38.72                     | 40.18                 | 0.9636635141861623
| ArithmeticTest.test_max2(uint256,uint256)                                     | 67.01                     | 69.38                 | 0.9658402997982128
| kontrol/src/tests/integration/test_foundry_prove.py::test_foundry_refute_node | 18.29                     | 17.52                 | 1.0439497716894977
| TOTAL                                                                         | 478.11000000000007        | 502.19                | 0.9520500209084213

some noise is introduced by `test_foundry_prove.py::test_foundry_refute_node` (the last item), which fails on both master and feature branch.